requirements.txt
confluent-kafka[avro,json]
python-dotenv